requirements.txt
streamlit
Pillow
PyPDF2
pdf2image